# docs/api/gateway-openapi.yaml
openapi: 3.0.3
info:
  title: NasibaShop API (via Kong Gateway)
  description: |
    Единая точка входа для клиентов. Базовый URL в локальной разработке: `http://localhost:8000`.
    Пути соответствуют маршрутизации Kong (`services/api-gateway/declarative/kong.yml`).
  version: 0.1.0
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Локальный Kong proxy
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Регистрация (email/пароль)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Создано
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Вход
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Токены
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление access-токена
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /api/auth/validate:
    post:
      tags: [Auth]
      summary: Проверка токена
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /api/users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [Users]
      summary: Профиль пользователя
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Пользователь
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Users]
      summary: Обновление профиля
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /api/users/{id}/addresses:
    post:
      tags: [Users]
      summary: Добавить адрес
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Создано
  /api/products:
    get:
      tags: [Products]
      summary: Список товаров
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: categoryId
          in: query
          schema: { type: string }
        - name: storeId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, example: ACTIVE }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Список и пагинация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      tags: [Products]
      summary: Создать товар
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "201":
          description: Создано
  /api/products/{id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [Products]
      summary: Товар с категорией и остатками
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: "#/components/schemas/ProductFull"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Products]
      summary: Обновить товар
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: OK
    delete:
      tags: [Products]
      summary: Удалить товар
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Удалено
  /api/products/{id}/stock:
    patch:
      tags: [Products]
      summary: Обновить остаток
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                variantId: { type: string }
                storeId: { type: string }
                quantity: { type: integer }
      responses:
        "200":
          description: OK
  /api/categories:
    get:
      tags: [Categories]
      summary: Дерево категорий
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
    post:
      tags: [Categories]
      summary: Создать категорию
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        "201":
          description: Создано
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 access token от user-service
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Не авторизован
    NotFound:
      description: Не найденo
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Category:
      type: object
      properties:
        id: { type: string }
        parentId: { type: string }
        slug: { type: string }
        nameUz: { type: string }
        sortOrder: { type: integer }
        children:
          type: array
          items:
            $ref: "#/components/schemas/Category"
    Product:
      type: object
      properties:
        id: { type: string }
        storeId: { type: string }
        categoryId: { type: string }
        titleUz: { type: string }
        descriptionUz: { type: string }
        brand: { type: string }
        priceUnits: { type: integer, format: int64 }
        currencyCode: { type: string }
        discountPercent: { type: integer }
        status: { type: string }
    ProductFull:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/Product"
        category:
          $ref: "#/components/schemas/Category"
        stocks:
          type: array
          items:
            type: object
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        pagination:
          type: object
          properties:
            totalCount:
              type: integer